src/requirements.txt
aiohttp==3.8.4
cerberus==1.3.4
cryptography==39.0.1
numpy==1.24.1
pandas==2.0.0
PyMySQL==1.0.2
python-crontab==2.7.1
requests==2.28.2
SQLAlchemy==2.0.3
mysql-connector-python==8.0.32
PyYAML==6.0